+++ src/web/modules/gateway_accounts/detail.njk
@@ -189,7 +189,32 @@
       </div>
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key"><span class="govuk-caption-m">MOTO</span></dt>
-        <dd class="govuk-summary-list__value">{{ 'Enabled' if account.allow_moto else 'Disabled' }}</dd>
+        <dd class="govuk-summary-list__value">
+          {% if account.allow_moto or
+            account.allow_authorisation_api or
+            account.allow_telephone_payment_notifications
+          %}
+            <ul class="govuk-list">
+              {% if account.allow_moto %}<li>Web MOTO payments enabled</li>{% endif %}
+              {% if account.allow_authorisation_api %}<li>Authorisation API enabled</li>{% endif %}
+              {% if account.allow_telephone_payment_notifications %}<li>Telephone payment notifications enabled</li>{% endif %}
+            </ul>
+          {% else %}
+            Disabled
+          {% endif %}
+        </dd>
+        <dd class="govuk-summary-list__actions">
+          <a class="govuk-link" href="/gateway_accounts/{{ gatewayAccountId }}/moto">Manage<span class="govuk-visually-hidden"> MOTO settings</span></a>
+        </dd>
+      </div>
+      <div class="govuk-summary-list__row">
+        <dt class="govuk-summary-list__key"><span class="govuk-caption-m">Agent-initiated MOTO link</span></dt>
+        <dd class="govuk-summary-list__value">
+          {{ 'Enabled' if motoPaymentLinkExists and services.agent_initiated_moto_enabled else 'Disabled' }}
+        </dd>
+        <dd class="govuk-summary-list__actions">
+          <a class="govuk-link" href="/gateway_accounts/{{ gatewayAccountId }}/agent_initiated_moto">Manage<span class="govuk-visually-hidden"> agent-initiated MOTO link</span></a>
+        </dd>
       </div>
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key"><span class="govuk-caption-m">Recurring enabled</span></dt>
@@ -338,46 +363,11 @@
     })
     }}
     {{ govukButton({
-      text: "Configure agent-initiated MOTO payments",
-      href: "/gateway_accounts/" + gatewayAccountId + "/agent_initiated_moto"
-    })
-    }}
-    {{ govukButton({
       text: "Switch PSP",
       href: "/gateway_accounts/" + gatewayAccountId + "/switch_psp"
     })
     }}
   </div>
-
-  {% if account.allow_moto === false %}
-    <div>
-      <h2 class="govuk-heading-s">Enable MOTO payments</h2>
-      <p class="govuk-body">Please ensure the service has passed PCI requirements before enabling MOTO payments.</p>
-      <p class="govuk-body">MOTO payments allow services to create payments that bypass security. MOTO payments have no 3DS checks and billing addresses aren't required. This is inherently insecure.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_moto_payments">
-        {{ govukButton({
-          text: "Enable MOTO payments",
-          classes: "govuk-button--warning"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
-    </div>
-  {% endif %}
-
-  {% if account.allow_moto === true %}
-    <div>
-      <h2 class="govuk-heading-s">MOTO payments are enabled</h2>
-      <p class="govuk-body">MOTO payments allow services to create payments that bypass security. MOTO payments have no 3DS checks and billing addresses aren't required. This is inherently insecure.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_moto_payments">
-        {{ govukButton({
-          text: "Disable MOTO payments"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
-    </div>
-  {% endif %}
 
   {% if account.block_prepaid_cards === true %}
     <div>
@@ -403,66 +393,6 @@
           href: "/gateway_accounts/" + gatewayAccountId + "/block_prepaid_cards/toggle"
           })
         }}
-    </div>
-  {% endif %}
-
-  {% if account.allow_telephone_payment_notifications === false %}
-    <div>
-      <h2 class="govuk-heading-s">Enable telephone payment notifications</h2>
-      <p class="govuk-body">Allows use of the <a class="govuk-link govuk-link--no-visited-state" href="https://github.com/alphagov/pay-telephone-payments">API to store records of telephone payments</a> which were taken outside of GOV.UK Pay.</p>
-      <p class="govuk-body">It should be ensured that the gateway account is only used for this purpose by the service and they have a separate service if they take payments using our payment pages.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_allow_telephone_payment_notifications">
-        {{ govukButton({
-          text: "Enable telephone payment notifications",
-          classes: "govuk-button--warning"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
-    </div>
-  {% endif %}
-
-  {% if account.allow_telephone_payment_notifications === true %}
-    <div>
-      <h2 class="govuk-heading-s">Telephone payment notifications are enabled</h2>
-      <p class="govuk-body">Use of the <a class="govuk-link govuk-link--no-visited-state" href="https://github.com/alphagov/pay-telephone-payments">telephone payment notifications API</a> is enabled for this account. This is used to store records of telephone payments which were taken outside of GOV.UK Pay.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_allow_telephone_payment_notifications">
-        {{ govukButton({
-          text: "Disable telephone payment notifications"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
-    </div>
-  {% endif %}
-
-    {% if account.allow_authorisation_api === false %}
-    <div>
-      <h2 class="govuk-heading-s">Use of the payment authorisation API is disabled</h2>
-      <p class="govuk-body">Allows creating MOTO payments with an authorisation_mode of ‘api‘. These payments are authorised by sending the card details in a POST request to Public API.</p>
-      <p class="govuk-body">Note that corporate card surcharges will not be applied to payments created in this way.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_allow_authorisation_api">
-        {{ govukButton({
-          text: "Enable use of the payment authorisation API",
-          classes: "govuk-button--warning"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
-    </div>
-  {% endif %}
-
-  {% if account.allow_authorisation_api === true %}
-    <div>
-      <h2 class="govuk-heading-s">Use of the payment authorisation API is enabled</h2>
-      <p class="govuk-body">Creating MOTO payments with an authorisation_mode of ‘api‘ is allowed for this account. These payments are authorised by sending the card details in a POST request to Public API.</p>
-      <form method="POST" action="/gateway_accounts/{{ gatewayAccountId }}/toggle_allow_authorisation_api">
-        {{ govukButton({
-          text: "Disable use of the payment authorisation API"
-          })
-        }}
-        <input type="hidden" name="_csrf" value="{{ csrf }}">
-      </form>
     </div>
   {% endif %}
