+++ src/web/modules/services/index.ts
@@ -10,7 +10,6 @@ import {
   searchRequest,
   addTestAccount,
   submitTestAccountProvider,
-  toggleAgentInitiatedMotoEnabledFlag,
   toggleArchiveService,
   toggleExperimentalFeaturesEnabledFlag,
   toggleTerminalStateRedirectFlag,
@@ -40,7 +39,6 @@ export default {
   searchRequest: searchRequest,
   toggleTerminalStateRedirectFlag: toggleTerminalStateRedirectFlag,
   toggleExperimentalFeaturesEnabled: toggleExperimentalFeaturesEnabledFlag,
-  toggleAgentInitiatedMotoEnabled: toggleAgentInitiatedMotoEnabledFlag,
   updateOrganisationForm: updateOrganisationForm,
   updateOrganisation: updateOrganisation,
   toggleArchiveService: toggleArchiveService,
