+++ src/web/router.js
@@ -68,8 +68,6 @@ router.get('/gateway_accounts/:id/surcharge', auth.secured(PermissionLevel.VIEW_
 router.post('/gateway_accounts/:id/surcharge', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.updateSurcharge)
 router.get('/gateway_accounts/:id/email_branding', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.emailBranding)
 router.post('/gateway_accounts/:id/email_branding', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.updateEmailBranding)
-router.post('/gateway_accounts/:id/toggle_moto_payments', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleMotoPayments)
-router.post('/gateway_accounts/:id/toggle_allow_telephone_payment_notifications', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleAllowTelephonePaymentNotifications)
 router.post('/gateway_accounts/:id/toggle_send_payer_ip_address_to_gateway', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleSendPayerIpAddressToGateway)
 router.post('/gateway_accounts/:id/toggle_send_payer_email_to_gateway', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleSendPayerEmailToGateway)
 router.post('/gateway_accounts/:id/toggle_worldpay_exemption_engine', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleWorldpayExemptionEngine)
@@ -77,16 +75,22 @@ router.post('/gateway_accounts/:id/toggle_send_reference_to_gateway', auth.secur
 router.get('/gateway_accounts/:id/disable', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.disableReasonPage)
 router.post('/gateway_accounts/:id/disable', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.disable)
 router.post('/gateway_accounts/:id/enable', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.enable)
-router.post('/gateway_accounts/:id/toggle_allow_authorisation_api', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleAllowAuthorisationApi)
 router.post('/gateway_accounts/:id/toggle_recurring_enabled', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleRecurringEnabled)
 router.get('/gateway_accounts/:id/stripe_statement_descriptor', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.updateStripeStatementDescriptorPage)
 router.post('/gateway_accounts/:id/stripe_statement_descriptor', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.updateStripeStatementDescriptor)
 router.get('/gateway_accounts/:id/stripe_payout_descriptor', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.updateStripePayoutDescriptorPage)
 router.post('/gateway_accounts/:id/stripe_payout_descriptor', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.updateStripePayoutDescriptor)
-router.get('/gateway_accounts/:id/agent_initiated_moto', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.agentInitiatedMotoPage)
-router.post('/gateway_accounts/:id/agent_initiated_moto', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.createAgentInitiatedMotoProduct)
 router.get('/gateway_accounts/:id/switch_psp', auth.secured(PermissionLevel.VIEW_ONLY), switchPSP.switchPSPPage)
 router.post('/gateway_accounts/:id/switch_psp', auth.secured(PermissionLevel.USER_SUPPORT), switchPSP.postSwitchPSP)
+
+router.get('/gateway_accounts/:id/moto', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.motoSettings)
+router.post('/gateway_accounts/:id/toggle_moto_payments', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleMotoPayments)
+router.get('/gateway_accounts/:id/agent_initiated_moto', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.agentInitiatedMoto)
+router.get('/gateway_accounts/:id/agent_initiated_moto_product', auth.secured(PermissionLevel.VIEW_ONLY), gatewayAccounts.agentInitiatedMotoProduct)
+router.post('/gateway_accounts/:id/agent_initiated_moto_product', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.createAgentInitiatedMotoProduct)
+router.get('/gateway_accounts/:id/toggle_agent_initiated_moto_enabled', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleAgentInitiatedMotoEnabledFlag)
+router.post('/gateway_accounts/:id/toggle_allow_authorisation_api', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleAllowAuthorisationApi)
+router.post('/gateway_accounts/:id/toggle_allow_telephone_payment_notifications', auth.secured(PermissionLevel.USER_SUPPORT), gatewayAccounts.toggleAllowTelephonePaymentNotifications)
 
 router.get('/gateway_accounts/:accountId/payment_links', auth.secured(PermissionLevel.VIEW_ONLY), paymentLinks.list)
 router.get('/gateway_accounts/:accountId/payment_links/csv', auth.secured(PermissionLevel.VIEW_ONLY), paymentLinks.listCSV)
@@ -103,7 +107,6 @@ router.get('/services/:id/link_accounts', auth.secured(PermissionLevel.ADMIN), s
 router.post('/services/:id/link_accounts', auth.secured(PermissionLevel.ADMIN), services.updateLinkAccounts.http, services.updateLinkAccounts.exceptions)
 router.get('/services/:id/toggle_terminal_state_redirect', auth.secured(PermissionLevel.USER_SUPPORT), services.toggleTerminalStateRedirectFlag)
 router.get('/services/:id/toggle_experimental_features_enabled', auth.secured(PermissionLevel.USER_SUPPORT), services.toggleExperimentalFeaturesEnabled)
-router.get('/services/:id/toggle_agent_initiated_moto_enabled', auth.secured(PermissionLevel.USER_SUPPORT), services.toggleAgentInitiatedMotoEnabled)
 router.get('/services/:id/toggle_archived_status', auth.secured(PermissionLevel.USER_SUPPORT), services.toggleArchiveService)
 router.get('/services/:id/organisation', auth.secured(PermissionLevel.USER_SUPPORT), services.updateOrganisationForm)
 router.post('/services/:id/organisation', auth.secured(PermissionLevel.USER_SUPPORT), services.updateOrganisation)
